+++ NEWS.md
@@ -1,3 +1,10 @@
+# colorist 0.1.1
+
+- Two vignettes added
+- Added an additional lambda parameter to map_single function to control visualization of specificity values
+- Added custom_hues argument to palette_set function
+- Bug fixes
+
 # colorist 0.1.0
 
 - First CRAN release